+++ src/components/StacSource.vue
@@ -1,6 +1,12 @@
 <template>
   <nav class="share">
     <b-button-group>
+      <b-button
+        v-if="canFavorite" size="sm" variant="outline-primary"
+        :title="favoriteTitle" :pressed="isFavorite" @click="toggleFavorite"
+      >
+        <b-icon-star-fill v-if="isFavorite" /><b-icon-star v-else /><span class="button-label">{{ $t('favorites.label') }}</span>
+      </b-button>
       <b-dropdown
         v-if="canManage" size="sm" variant="outline-primary" tabindex="0"
       >
@@ -107,6 +113,7 @@
 import { mapGetters, mapState } from 'vuex';
 import { defineAsyncComponent } from 'vue';
 import { BDropdown, BDropdownItem } from 'bootstrap-vue-next';
+import { STAC } from 'stac-js';
 
 import Url from './Url.vue';
 import CopyButton from './CopyButton.vue';
@@ -139,9 +146,18 @@ export default {
     };
   },
   computed: {
-    ...mapState(['data', 'socialSharing', 'url']),
+    ...mapState(['data', 'showFavorites', 'socialSharing', 'url']),
     ...mapGetters(['toBrowserPath', 'collectionLink', 'parentLink', 'rootLink']),
     ...mapGetters('manager', ['browserPaths', 'canEdit', 'canDelete', 'canManage', 'canAddCollections', 'canAddItems', 'externalCreateLink', 'externalEditLink']),
+    canFavorite() {
+      return this.showFavorites && this.$route.name === 'browse' && this.data instanceof STAC;
+    },
+    isFavorite() {
+      return this.$store.getters['favorites/isFavorite'](this.data);
+    },
+    favoriteTitle() {
+      return this.isFavorite ? this.$t('favorites.remove') : this.$t('favorites.add');
+    },
     stacVersion() {
       return this.data?.stac_version;
     },
@@ -159,6 +175,9 @@ export default {
   methods: {
     browserUrl() {
       return window.location.toString();
+    },
+    async toggleFavorite() {
+      await this.$store.dispatch('favorites/toggle', this.data);
     },
     async deleteThis() {
       this.deleting = true;